v0.3.4-alpha ⬇︎
The ContactForm handles:
errors for empty fields
email missing @ and .
Once everything is good, during the submit, it makes a fetch, at the recently added serve library route ( check the related commits ), and show a "message sent" message.

small changes in the css styling.

not the "very ready" but it is a good start ;)