v0.3.4-alpha ⬇︎
It initializes the server

the imported route [message]
The route [message] will receive the fetch request from the ContactForm components and send a response {message:received} to the React Application

It calls the module routing that handles the requests and responses through the EventEmitter object. It listens for the events emitted from the dispatcher module

It sets the dispatcher modules that is responsible to send files to the browser, calling the routes registered and handling errors.
It emits:

[error] in case of errors

[read] when the file has be read and the buffer is emitted, Ready for the server.
[ready] now everything is set and the routing can send the buffer to the browser and the logger print to terminal the connection infos.

[serve] same as [read] but when the API interface of the serve library is called.
[api] same as [ready] but when the API interface of the serve library is called.